refactor: 채팅 컴포넌트 디렉토리를 기능 단위로 재구성
flat 구조의 components/를 chat-box, chat-list, sidebar, join-request로 그룹핑하여
컴포넌트 간 의존 관계를 디렉토리 구조로 명확하게 표현.
