+++ apps/frontend/src/app/chat/[id]/page.tsx
@@ -1,4 +1,4 @@
-import ChatBox from '../components/ChatBox';
+import ChatBox from '../components/chat-box/ChatBox';
 
 interface ChatRoomPageProps {
   params: {
