+++ apps/frontend/src/app/chat/components/chat-box/ChatInput.tsx
