+++ apps/frontend/src/app/chat/components/chat-box/MessageBubble.tsx
