+++ apps/frontend/src/app/chat/components/join-request/JoinRequestDropdown.tsx
