+++ apps/frontend/src/app/chat/components/sidebar/ChatInfoSidebar.tsx
