+++ apps/frontend/src/app/chat/components/sidebar/KickConfirmDialog.tsx
